requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
pdfplumber==0.10.2
python-docx==0.8.11
python-pptx==0.6.21
pandas==2.1.0
openpyxl==3.1.2
reportlab==4.0.4
werkzeug==2.3.7
